# openapi.yaml
...
openapi: 3.1.0
info:
  title: Issues API Testbed
  version: 0.1.0
  summary: A validated wrapper over the GitHub Issues REST API, with signed webhook ingestion.
  description: |
    This service exposes its own HTTP API for issue CRUD and comments on a single
    GitHub repository, and ingests `issues` / `issue_comment` webhooks from that
    repository.

    ### Authentication
    Clients of *this* service do not authenticate. The GitHub credential is held
    server-side in the `GITHUB_TOKEN` environment variable and is never accepted
    from, or echoed to, a caller. The `githubToken` security scheme below documents
    that server-to-GitHub relationship so the dependency is visible in the contract.

    `POST /webhook` is authenticated differently: GitHub signs each delivery with
    `WEBHOOK_SECRET`, and the service verifies the `X-Hub-Signature-256` header with
    a constant-time comparison over the raw request body.

    ### Errors
    Every failure — validation, upstream, or unexpected — uses the same `Error`
    envelope, so a client parses one shape. `details` is free-form and carries the
    upstream context when there is any.
  contact:
    name: Anita Agasaveeran
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8000"
...
paths:
...
  /webhook:
    post:
      tags: [webhooks]
      operationId: receiveWebhook
      summary: Receive a GitHub webhook delivery
      description: |
        Verifies the HMAC-SHA256 signature over the **raw** request body, records
        the delivery, and acknowledges immediately. Slow work is deferred to a
        background task so the acknowledgement is never blocked.

        Processing is idempotent: the store keys on `(X-GitHub-Delivery, action)`,
        so a redelivery of the same event is a no-op that still returns `204`.

        Checks run signature-first on purpose — an unauthenticated caller should not
        be able to discover which events the service accepts.
      security:
        - webhookSignature: []
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          description: Event name. Only `issues`, `issue_comment`, and `ping` are accepted.
          schema:
            type: string
            enum: [issues, issue_comment, ping]
          example: issues
        - name: X-GitHub-Delivery
          in: header
          required: true
          description: Unique delivery id; reused by GitHub across redeliveries.
          schema:
            type: string
          example: 72d3162e-cc78-11e3-81ab-4c9367dc0958
        - name: X-Hub-Signature-256
          in: header
          required: true
          description: "`sha256=` followed by the HMAC-SHA256 hex digest of the raw body."
          schema:
            type: string
            pattern: "^sha256=[0-9a-f]{64}$"
      requestBody:
        required: true
        description: The raw GitHub event payload, passed through unmodified.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
            examples:
              issueOpened:
                summary: issues / opened
                value:
                  action: opened
                  issue:
                    number: 42
                    title: Checkout button is dead
                    state: open
                  sender:
                    login: anita-agasaveeran
              commentCreated:
                summary: issue_comment / created
                value:
                  action: created
                  issue:
                    number: 42
                  comment:
                    id: 3100000001
                    body: Reproduced on Safari 17.
                  sender:
                    login: anita-agasaveeran
              ping:
                summary: ping
                value:
                  zen: Non-blocking is better than blocking.
                  hook_id: 500000001
      responses:
        "204":
          description: |
            Delivery accepted. Returned for a first delivery and for a redelivery
            alike — the caller cannot distinguish them, and does not need to.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "400":
          description: Unsupported event, missing header, or malformed payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unsupportedEvent:
                  summary: Unsupported event
                  value:
                    error:
                      code: validation_error
                      message: "Unsupported webhook event 'push'."
                      details:
                        event: push
                        supported: [issue_comment, issues, ping]
                      request_id: 5ca4ba68-d427-4c76-af34-ed5d9b23e01b
                missingAction:
                  summary: Payload without an action
                  value:
                    error:
                      code: validation_error
                      message: "Event 'issues' requires an 'action' field."
                      details:
                        event: issues
                        reason: action_missing
                      request_id: 5ca4ba68-d427-4c76-af34-ed5d9b23e01b
        "401":
          description: |
            Missing or invalid signature. The message is identical for every cause —
            absent header, wrong digest, tampered body — so it cannot be used to
            iterate towards a forgery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalidSignature:
                  summary: Signature mismatch
                  value:
                    error:
                      code: unauthorized
                      message: Missing or invalid webhook signature.
                      details:
                        reason: signature_mismatch
                      request_id: b00d0e4c-46da-4bc3-8ec6-66d22f349b45
  /events:
    get:
      tags: [webhooks]
      operationId: listEvents
      summary: List recently processed webhook deliveries
      description: |
        Returns the most recent deliveries, newest first — a debugging aid for
        confirming that a webhook arrived and was recorded. Raw payloads are
        retained in the local store but deliberately excluded from this response.
      security: []
      parameters:
        - name: limit
          in: query
          required: false
          description: How many deliveries to return. Also capped by `EVENTS_MAX_LIMIT`.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: Recent deliveries, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventRecord"
              examples:
                recent:
                  summary: Two deliveries
                  value:
                    - id: 72d3162e-cc78-11e3-81ab-4c9367dc0958
                      event: issue_comment
                      action: created
                      issue_number: 42
                      timestamp: "2026-09-02T12:00:03.114221Z"
                    - id: 5f8a1c40-cc78-11e3-81ab-4c9367dc0958
                      event: issues
                      action: opened
                      issue_number: 42
                      timestamp: "2026-09-01T10:00:01.882004Z"
                empty:
                  summary: Nothing received yet
                  value: []
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub personal access token
      description: |
        The token this service presents **to GitHub**, read from `GITHUB_TOKEN` at
        startup. It is never supplied by a client of this API and never appears in
        a response or a log line. Requires `Issues: read & write` and
        `Metadata: read` on the configured repository.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: |
        HMAC-SHA256 of the raw request body, keyed with `WEBHOOK_SECRET` and sent by
        GitHub as `sha256=<hex digest>`. Verified with a constant-time comparison.
...
  headers:
...
    RequestId:
      description: Correlation id, echoed from the request or generated per request.
      schema:
        type: string
      example: b00d0e4c-46da-4bc3-8ec6-66d22f349b45
  schemas:
...
    EventRecord:
      type: object
      required: [id, event, timestamp]
      description: One processed webhook delivery.
      properties:
        id:
          type: string
          description: The `X-GitHub-Delivery` id.
        event:
          type: string
          enum: [issues, issue_comment, ping]
        action:
          type: [string, "null"]
          description: Null for `ping`, which carries no action.
          examples: [opened]
        issue_number:
          type: [integer, "null"]
        timestamp:
          type: string
          format: date-time
          description: When this service recorded the delivery, in UTC.
    WebhookPayload:
      type: object
      description: |
        A GitHub event payload. Only the indexed fields are described here; the rest
        of the object is accepted and stored verbatim.
      properties:
        action:
          type: string
          description: Required for `issues` and `issue_comment`; absent on `ping`.
        issue:
          type: object
          properties:
            number:
              type: integer
        comment:
          type: object
          properties:
            id:
              type: integer
        sender:
          type: object
          properties:
            login:
              type: string
      additionalProperties: true
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          enum:
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - rate_limited
            - upstream_error
            - upstream_unavailable
            - internal_error
            - http_error
        message:
          type: string
          description: Human-readable explanation, safe to surface to an operator.
        details:
          description: |
            Free-form context. Carries `upstream_status` and `upstream_message` when
            the failure came from GitHub. Never contains secrets or signatures.
        request_id:
          type: [string, "null"]
          description: Correlation id, matching the `X-Request-ID` response header.
    Error:
      type: object
      required: [error]
      description: The single error envelope used by every failing response.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingTitle:
              summary: Missing required field
              value:
                error:
                  code: validation_error
                  message: Request payload or query parameters are invalid.
                  details:
                    - location: [body, title]
                      message: Field required
                      type: missing
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
            invalidState:
              summary: Unsupported state filter
              value:
                error:
                  code: validation_error
                  message: Request payload or query parameters are invalid.
                  details:
                    - location: [query, state]
                      message: "Input should be 'open', 'closed' or 'all'"
                      type: enum
                  request_id: 7830098e-2558-4aae-abb2-08277b81ec98
...